Adds `yesod-simple-rio` template
CF: yesodweb/yesod#1602
CC: @snoyberg

This template takes `yesod-simple` and replaces `classy-prelude-*`
with `rio`. I have checked that everything builds and runs out of the
box, and that the tests are green.

I have not (yet) used it in anger to develop an app, and I am but a
Yesod and RIO neophyte, so it needs careful reviewing.

Some points:

  * bounds on `aeson` and `yaml` were relaxed to allow those in the
    latest lts release (13.25);
  * `http-types`, `yesod-newsfeed` and `persistent` was added with
    no bounds restrictions &mdash; what restrictions should we use
    on these packages?
  * `rio >= 0.1.8.0` : is my starting point &mdash; can we relax the
    upper bound? What upper bound should we use?
  * To get `rio` to work with the current stack I am hiding parts of the
    `rio` logging system (I guess these will be addressed by
    [this work](https://github.com/yesodweb/yesod/issues/1602#issuecomment-496503671)).